Move all cases to ddl/tests
