+++ ddl/tests/adminpause/BUILD.bazel
@@ -1,13 +1,13 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "adminpausetest",
+    name = "adminpause",
     srcs = [
         "ddl_data_generation.go",
         "ddl_stmt_cases.go",
         "global.go",
     ],
-    importpath = "github.com/pingcap/tidb/ddl/adminpausetest",
+    importpath = "github.com/pingcap/tidb/ddl/tests/adminpause",
     visibility = ["//visibility:public"],
     deps = [
         "//parser/model",
@@ -17,7 +17,7 @@ go_library(
 )
 
 go_test(
-    name = "adminpausetest_test",
+    name = "adminpause_test",
     timeout = "short",
     srcs = [
         "main_test.go",
@@ -25,7 +25,7 @@ go_test(
         "pause_negative_test.go",
         "pause_resume_test.go",
     ],
-    embed = [":adminpausetest"],
+    embed = [":adminpause"],
     flaky = True,
     race = "on",
     shard_count = 5,
