+++ ddl/tests/adminpause/ddl_data_generation.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package adminpausetest
+package adminpause
 
 import (
 	"fmt"
