+++ ddl/tests/adminpause/ddl_stmt_cases.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package adminpausetest
+package adminpause
 
 import (
 	"sync"
